# swagger.yaml
openapi: 3.0.0
info:
  title: Cart API
  description: API for managing user shopping carts
  version: 1.0.0
servers:
  - url: http://localhost:8080/
  - url: https://cartserviceem.azurewebsites.net/
paths:
  /cart:
    get:
      summary: Retrieve all items in the cart
      security:
        - JWT: []  # Apply JWT authentication
      responses:
        '200':
          description: A list of items in the cart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItem'
    delete:
      summary: Remove all items from the cart
      security:
        - JWT: []  # Apply JWT authentication
      responses:
        '204':
          description: All items successfully removed from the cart
        '401':
          description: Access token is missing or invalid

    post:
      summary: Add a new item to the cart
      security:
        - JWT: []  # Apply JWT authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItem'
      responses:
        '201':
          description: Item successfully added to the cart
        '401':
          description: Access token is missing or invalid
  /cart/{id}:
    patch:
      summary: Update an existing item in the cart
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the item to update
      security:
        - JWT: []  # Apply JWT authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItem'
      responses:
        '200':
          description: Item updated successfully
        '404':
          description: Item not found
        '401':
          description: Access token is missing or invalid
    delete:
      summary: Remove an item from the cart
      security:
        - JWT: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the item to remove
      responses:
        '204':
          description: Item successfully removed from the cart
        '404':
          description: Item not found
        '401':
          description: Access token is missing or invalid

components:
  schemas:
    CartItem:
      type: object
      properties:
        product:
          type: string
          description: The id of the product
        quantity:
          type: integer
          description: The quantity of the product in the cart
          minimum: 1
        price:
          type: number
          format: float
          description: The price of the product
      required:
        - product
        - quantity
        - price
  securitySchemes:
    JWT:            # arbitrary name for the security scheme
      type: http
      scheme: bearer